+++ src/Variables/DefaultVariables.jl
@@ -1,9 +1,16 @@
 
+export ContinuousScalar
 export ContinuousEuclid
+export Circular, Circle
 
 
-Base.convert(::Type{<:Tuple}, mani::ManifoldsBase.Manifold) = getManifolds(mani)
+Base.convert(::Type{<:Tuple}, ::InstanceType{Manifolds.Euclidean{Tuple{N}, ℝ}} ) where N = tuple([:Euclid for i in 1:N]...)
+Base.convert(::Type{<:Tuple}, ::InstanceType{Manifolds.Circle{ℝ}})  = (:Circular,)
 
+# Base.convert(::Type{<:Tuple}, mani::ManifoldsBase.Manifold) = getManifolds(mani)
+
+
+## Euclid 1
 
 """
 $(TYPEDEF)
@@ -13,24 +20,10 @@ Most basic continuous scalar variable in a `::DFG.AbstractDFG` object.
 DevNotes
 - TODO Consolidate with ContinuousEuclid{1}
 """
-@defVariable ContinuousScalar 1 (:Euclid,)
+@defVariable ContinuousScalar Euclidean(1) #1 (:Euclid,)
 
-"""
-$(TYPEDEF)
 
-Continuous variable of dimension `.dims` on manifold `.manifolds`.
-"""
-struct ContinuousMultivariate{T1 <: Tuple} <: InferenceVariable
-  dims::Int
-  manifolds::T1
-end
-
-function ContinuousMultivariate(x::Int;
-                                manifolds::T1=convert(Tuple, Euclidean(1))  )  where {T1 <: Tuple}
-  #
-  maniT = length(manifolds) < x ? ([manifolds[1] for i in 1:x]...,) : manifolds
-  ContinuousMultivariate{typeof(maniT)}(x, maniT)
-end
+## Euclid N
 
 
 """
@@ -41,10 +34,31 @@ struct ContinuousEuclid{N} <: InferenceVariable end
 
 ContinuousEuclid(x::Int) = ContinuousEuclid{x}()
 
-getDimension(::ContinuousEuclid{N}) where N = N::Int
-getManifolds(::ContinuousEuclid{N}) where N = ntuple(i -> :Euclid, N)
+getManifold(::Type{<:ContinuousEuclid{N}}) where N = Euclidean(N)                               # ntuple(i -> :Euclid, N)
+getDimension(val::Type{<:ContinuousEuclid{N}}) where N = manifold_dimension(getManifold(val))   # N::Int
+getManifolds(val::Type{<:ContinuousEuclid{N}}) where N = convert(Tuple, getManifold(val))       # ntuple(i -> :Euclid, N)
 
+getManifold(::ContinuousEuclid{N}) where N = Euclidean(N)                               # ntuple(i -> :Euclid, N)
+getDimension(val::ContinuousEuclid{N}) where N = manifold_dimension(getManifold(val))   # N::Int
+getManifolds(val::ContinuousEuclid{N}) where N = convert(Tuple, getManifold(val))       # ntuple(i -> :Euclid, N)
 
 Base.convert(::Type{<:ManifoldsBase.Manifold}, ::InstanceType{ContinuousEuclid{N}}) where N = Manifolds.Euclidean(N)
 Base.convert(::Type{<:ManifoldsBase.Manifold}, ::InstanceType{ContinuousScalar})    = Manifolds.Euclidean(1)
 
+
+## Circular
+
+
+"""
+$(TYPEDEF)
+
+Circular is a `Manifolds.Circle{ℝ}` mechanization of one rotation, with `theta in [-pi,pi)`.
+"""
+@defVariable Circular Circle()
+
+
+Base.convert(::Type{<:ManifoldsBase.Manifold}, ::InstanceType{Circular}) = Manifolds.Circle()
+
+
+
+#