+++ NEWS.md
@@ -11,6 +11,12 @@ Alternatively, either use the Github Blame, or the Github `/compare/v0.18.0...v0
 
 The list below highlights major breaking changes, and please note that significant efforts are made to properly deprecate old code/APIs according to normal semver workflow -- i.e. breaking changes go through at least one deprecatation (via warnings) on the dominant number in the version number.  E.g. v0.18 -> v0.19 (warnings) -> v0.20 (breaking).
 
+# Major changes in IIF v0.23
+
+- New `@defVariable` only uses `ManifoldsBase.Manifold` as base abstraction for variable types.
+# Major changes in IIF v0.22
+
+- Work in progress toward `ManifoldsBase.Manifold` as base abstraction for variable types.
 # Major changes in IIF v0.21
 
 - `CalcResidual` no longer takes a `residual` as input parameter and should return `residual`, see #467 .